+++ CHANGELOG.md
@@ -29,6 +29,7 @@ nav_order: 1
 ## [4.9.4] - 2023-12-13
 
 - Fix race issues with `aiven_mirrormaker_replication_flow` on create/update/delete operations   
+- Add `tech_emails` to services
 
 ## [4.9.3] - 2023-10-27