+++ CHANGELOG.md
@@ -26,6 +26,7 @@ nav_order: 1
 - Add support for the `aiven_organization_user_group_member` resource, allowing the association of groups with the users. Please note that this resource is in the beta stage, and to use it, you would need to set the environment variable PROVIDER_AIVEN_ENABLE_BETA to a non-zero value.  
 - Retry kafka topic creation error `Cluster only has N broker(s), cannot set replication factor to M`
 - Fix Kafka Topic migration issues from V3 to V4.
+- Fix V3 to V4 migration issue related to cloud_name diff.
 
 ## [4.9.4] - 2023-12-13